+++ baseapp/iostats_nop.go
@@ -5,5 +5,4 @@ package baseapp
 import "github.com/line/ostracon/libs/log"
 
 func logIoStats(logger log.Logger) {
-	return
 }