+++ x/simulation/util.go
@@ -5,6 +5,13 @@ import (
 	"fmt"
 	"math/rand"
 	"testing"
+
+	"github.com/line/lbm-sdk/baseapp"
+	"github.com/line/lbm-sdk/client"
+	"github.com/line/lbm-sdk/codec"
+	"github.com/line/lbm-sdk/simapp/helpers"
+	"github.com/line/lbm-sdk/types"
+	simtype "github.com/line/lbm-sdk/types/simulation"
 )
 
 func getTestingMode(tb testing.TB) (testingMode bool, t *testing.T, b *testing.B) {
@@ -52,4 +59,67 @@ func mustMarshalJSONIndent(o interface{}) []byte {
 	}
 
 	return bz
+}
+
+// OperationInput is a struct that holds all the needed values to generate a tx and deliver it
+type OperationInput struct {
+	R               *rand.Rand
+	App             *baseapp.BaseApp
+	TxGen           client.TxConfig
+	Cdc             *codec.ProtoCodec
+	Msg             types.Msg
+	MsgType         string
+	CoinsSpentInMsg types.Coins
+	Context         types.Context
+	SimAccount      simtype.Account
+	AccountKeeper   AccountKeeper
+	Bankkeeper      BankKeeper
+	ModuleName      string
+}
+
+// GenAndDeliverTxWithRandFees generates a transaction with a random fee and delivers it.
+func GenAndDeliverTxWithRandFees(txCtx OperationInput) (simtype.OperationMsg, []simtype.FutureOperation, error) {
+	account := txCtx.AccountKeeper.GetAccount(txCtx.Context, txCtx.SimAccount.Address)
+	spendable := txCtx.Bankkeeper.SpendableCoins(txCtx.Context, account.GetAddress())
+
+	var fees types.Coins
+	var err error
+
+	coins, hasNeg := spendable.SafeSub(txCtx.CoinsSpentInMsg)
+	if hasNeg {
+		return simtype.NoOpMsg(txCtx.ModuleName, txCtx.MsgType, "message doesn't leave room for fees"), nil, err
+	}
+
+	fees, err = simtype.RandomFees(txCtx.R, txCtx.Context, coins)
+	if err != nil {
+		return simtype.NoOpMsg(txCtx.ModuleName, txCtx.MsgType, "unable to generate fees"), nil, err
+	}
+	return GenAndDeliverTx(txCtx, fees)
+}
+
+// GenAndDeliverTx generates a transactions and delivers it.
+func GenAndDeliverTx(txCtx OperationInput, fees types.Coins) (simtype.OperationMsg, []simtype.FutureOperation, error) {
+	account := txCtx.AccountKeeper.GetAccount(txCtx.Context, txCtx.SimAccount.Address)
+	tx, err := helpers.GenTx(
+		txCtx.TxGen,
+		[]types.Msg{txCtx.Msg},
+		fees,
+		helpers.DefaultGenTxGas,
+		txCtx.Context.ChainID(),
+		[]uint64{0},
+		[]uint64{account.GetSequence()},
+		txCtx.SimAccount.PrivKey,
+	)
+
+	if err != nil {
+		return simtype.NoOpMsg(txCtx.ModuleName, txCtx.MsgType, "unable to generate mock tx"), nil, err
+	}
+
+	_, _, err = txCtx.App.Deliver(txCtx.TxGen.TxEncoder(), tx)
+	if err != nil {
+		return simtype.NoOpMsg(txCtx.ModuleName, txCtx.MsgType, "unable to deliver tx"), nil, err
+	}
+
+	return simtype.NewOperationMsg(txCtx.Msg, true, ""), nil, nil
+
 }